feat: 랭킹 기능 레디스로 이전 (#63)
* feat: 랭킹 데이터 Redis로 이전

* refactor: GameRank 삭제, RankingResponse로 대체

* feat: 회원 게임 전적, 평균 타수, 평균 정확도 계산 로직 및 반환 응답값 추가

---------

+++ api/src/main/java/com/pgms/api/domain/game/dto/response/GameRoomMemberGetResponse.java
@@ -5,10 +5,10 @@ import com.pgms.coredomain.domain.game.GameRoomMember;
 public record GameRoomMemberGetResponse(
 	Long memberId,
 	String nickname,
-	int ranking,
+	Long ranking,
 	boolean readyStatus
 ) {
-	public static GameRoomMemberGetResponse from(GameRoomMember gameRoomMember, int ranking) {
+	public static GameRoomMemberGetResponse from(GameRoomMember gameRoomMember, Long ranking) {
 		return new GameRoomMemberGetResponse(
 			gameRoomMember.getMemberId(),
 			gameRoomMember.getNickname(),
